backend/app/requirements.txt
# FastAPI and related packages
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.7
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data processing and ML packages
joblib>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.2
xgboost>=1.7.5

# API and data fetching
requests>=2.28.0
python-dotenv>=1.0.0
yfinance>=0.2.18

# Caching and rate limiting
redis>=4.5.5

# Logging and monitoring
pythonjsonlogger>=2.0.7

# Optional dependencies
# transformers>=4.30.0  # For NLP models
# torch>=2.0.0  # For PyTorch models
# tensorflow>=2.12.0  # For TensorFlow models
